[#875] rename missed res packages
