+++ res/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.oztechan.ccc.resources" />
+<manifest package="com.oztechan.ccc.res" />
